# improve erf performance