--- a/scipy/special/cephes/ndtr.c
+++ b/scipy/special/cephes/ndtr.c
@@ -147,6 +147,7 @@
 #include <float.h>		/* DBL_EPSILON */
 #include "mconf.h"
 
+
 extern double SQRTH;
 extern double MAXLOG;