as suggested by @pv allow inlining polynomials, combined with unrolling and isnan replace this bumps performance quite considerably:

```
import numpy as np
import scipy.special as sps
r=np.random.rand(1e5)
%timeit -n 1000 -r 5 sps.erf(r)
%timeit -n 1000 -r 5 sps.erf(r + 2.)
```

before:

```
1000 loops, best of 5: 2.7 ms per loop
1000 loops, best of 5: 9.56 ms per loop
```

after

```
1000 loops, best of 5: 1.17 ms per loop
1000 loops, best of 5: 6.42 ms per loop
```