@pv 2015-03-21
Thanks, merged. Also moved the unroll+builtin-isnan part to apply to whole Cephes.