@rgommers 2022-08-07
I ran into this when cleaning up `_c99compat.h` (see gh-16800), and tested whether this still helps - it indeed does, so I didn't touch it:

```
In [2]: import numpy as np
   ...: import scipy.special as sps
   ...: r=np.random.rand(int(1e5))
   ...: %timeit -n 1000 -r 5 sps.erf(r)
   ...: %timeit -n 1000 -r 5 sps.erf(r + 2.)
395 µs ± 20.3 µs per loop (mean ± std. dev. of 5 runs, 1,000 loops each)
1.74 ms ± 31.1 µs per loop (mean ± std. dev. of 5 runs, 1,000 loops each)
```
remove the unroll pragma:
```
517 µs ± 7.35 µs per loop (mean ± std. dev. of 5 runs, 1,000 loops each)
1.87 ms ± 25.7 µs per loop (mean ± std. dev. of 5 runs, 1,000 loops each)
```
don't use `__builtins_isnan` at all:
```
723 µs ± 4.76 µs per loop (mean ± std. dev. of 5 runs, 1,000 loops each)
2.04 ms ± 7.47 µs per loop (mean ± std. dev. of 5 runs, 1,000 loops each)
```